move PeerAddress and NodeID into address.go
